Improvement to status message handling
git-svn-id: http://miranda.googlecode.com/svn/trunk@6022 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/MSN/msn_global.h
@@ -174,7 +174,7 @@ void  MSN_ShowPopup( const HANDLE hContact, const TCHAR* msg, int flags );
 
 LONG		MSN_SendPacket( HANDLE, const char* cmd, const char* params, ... );
 char*		MirandaStatusToMSN( int status );
-int			MSNStatusToMiranda( const char* status );
+WORD		MSNStatusToMiranda( const char* status );
 void        HtmlDecode( char* str );
 char*       HtmlEncode( const char* str );
 WCHAR*      HtmlEncodeW( const WCHAR* str );
